chore: moved start/demos to root
- show demos from both core toky and css-selector-parser
