+++ demos/css-selector-parser.html
@@ -1,0 +1,28 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
+    <title>Tokens</title>
+    <link rel="stylesheet" href="./css/style.css" />
+    <script type="importmap">
+      {
+        "imports": {
+          "toky": "/packages/toky/src/index.ts"
+        }
+      }
+    </script>
+  </head>
+  <body>
+    <main>
+      <textarea id="input">.x</textarea>
+      <pre id="output"></pre>
+    </main>
+    <script type="module">
+      import { initPreview, renderNestedToken } from './client/view.ts';
+      import { tokenizeSelector } from '../packages/css-selector-parser/src/index.ts';
+      initPreview(input, output, tokenizeSelector, renderNestedToken);
+    </script>
+  </body>
+</html>
