+++ demos/css-value-tokenizer.html
@@ -4,18 +4,18 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <title>Tokens</title>
+    <title>CSS Value Tokenizer</title>
     <link rel="stylesheet" href="./css/style.css" />
   </head>
   <body>
     <main>
-      <textarea id="input">.x</textarea>
+      <textarea id="input">10px value(x) 20px</textarea>
       <pre id="output"></pre>
     </main>
     <script type="module">
-      import { initPreview, renderNestedToken } from "./client/view.ts";
-      import { tokenizeSelector } from "../src/parsers/selector-tokenizer.ts";
-      initPreview(input, output, tokenizeSelector, renderNestedToken);
+      import { initPreview } from './client/view.ts';
+      import { createCssValueAST } from '../packages/toky/src/parsers/css-value-tokenizer.ts';
+      initPreview(input, output, createCssValueAST);
     </script>
   </body>
 </html>
