+++ demos/import-tokenizer.html
@@ -1,36 +1,30 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
+    <title>Tokens</title>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <title>Tokens</title>
     <link rel="stylesheet" href="./css/style.css" />
   </head>
   <body>
     <main>
       <textarea id="input">
-@function getLegacyAssetUrl($assetName) {
-    @return url('../../legacy-assets/' + $assetName);
-}
-
-[data-x="url('asdasda')"]         
-
-/* url('asd') */ 
-
-.x {
-    bg: url(../x) ,url('../y');
-    color: var(--x, url('../z'))
-
-}
+import * from "1";
+import Name from "2";
+import * as Name from "3";
+import * , {name, name1, name2 as name3} from "4";
+import {name} from "5";
+import *, {name as name1} from "6";
+import "7";
         </textarea
       >
       <pre id="output"></pre>
     </main>
     <script type="module">
-      import { initPreview } from "./client/view.ts";
-      import { tokenizeCSSUrls } from "../src/parsers/url-tokenizer.ts";
-      initPreview(input, output, tokenizeCSSUrls);
+      import { initPreview } from './client/view.ts';
+      import { tokenizeImports } from '../packages/toky/src/parsers/import-tokenizer.ts';
+      initPreview(input, output, tokenizeImports);
     </script>
   </body>
 </html>
