+++ demos/js-tokenizer.html
@@ -1,21 +1,21 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
+    <title>Tokens</title>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <title>CSS Value Tokenizer</title>
     <link rel="stylesheet" href="./css/style.css" />
   </head>
   <body>
     <main>
-      <textarea id="input">10px value(x) 20px</textarea>
+      <textarea id="input"></textarea>
       <pre id="output"></pre>
     </main>
     <script type="module">
-      import { initPreview } from "./client/view.ts";
-      import { createCssValueAST } from "../src/parsers/css-value-tokenizer.ts";
-      initPreview(input, output, createCssValueAST);
+      import { initPreview } from './client/view.ts';
+      import { tokenizeJS } from '../packages/toky/src/parsers/js-tokenizer.ts';
+      initPreview(input, output, tokenizeJS);
     </script>
   </body>
 </html>
