--- packages/toky/demos/js-tokenizer.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <title>Tokens</title>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <link rel="stylesheet" href="./css/style.css" />
-  </head>
-  <body>
-    <main>
-      <textarea id="input"></textarea>
-      <pre id="output"></pre>
-    </main>
-    <script type="module">
-      import { initPreview } from "./client/view.ts";
-      import { tokenizeJS } from "../src/parsers/js-tokenizer.ts";
-      initPreview(input, output, tokenizeJS);
-    </script>
-  </body>
-</html>
