+++ packages/toky/package.json
@@ -7,8 +7,7 @@
   "scripts": {
     "build": "rimraf dist && tsc",
     "prepack": "npm run build",
-    "test": "mocha \"./dist/test/**/*.spec.js\"",
-    "start": "node scripts/dev-server.js"
+    "test": "mocha \"./dist/test/**/*.spec.js\""
   },
   "keywords": [
     "parser",
